Make `crypto-chart` example work as a Now static deployment
This change utilizes static builds by adding a Dockerfile that
builds the static assets into the `/public` directory inside
the docker image.

Also, the `extends` clause in the `tsconfig.json` file was not
working since the include file is actually not present within the
deployment. Quick fix is to just copy+paste the contents of that
file into this `tsconfig.json` file to do the "extend" manually.